Add: ability to regenerate cache translation

+++ app/Plugin/Locale/View/Translations/admin_list.ctp
@@ -7,8 +7,9 @@ $tSettings = array(
         ),
         __t('Actions') => array(
             'value' => "
-                <a href='{url}/admin/locale/translations/edit/{Translation.id}{/url}'>" . __t('edit') . "</a>
-                <a href='{url}/admin/locale/translations/delete/{Translation.id}{/url}'>" . __t('delete') . "</a>
+                <a href='{url}/admin/locale/translations/edit/{Translation.id}{/url}'>" . __t('edit') . "</a> |
+                <a href='{url}/admin/locale/translations/regenerate/{Translation.id}{/url}' title='" . __t('Regenerate translation cache') . "'>" . __t('regenerate') . "</a> |
+                <a href='{url}/admin/locale/translations/delete/{Translation.id}{/url}' onclick='return confirm(\"" . __t('Delete this entry ?') . "\");'>" . __t('delete') . "</a>
             ",
             'thOptions' => array('align' => 'right'),
             'tdOptions' => array('align' => 'right')
@@ -17,13 +18,11 @@ $tSettings = array(
     'noItemsMessage' => __t('There are no translations to display'),
     'paginate' => true,
     'headerPosition' => 'top',
-    'tableOptions' => array('width' => '100%')    # table attributes
+    'tableOptions' => array('width' => '100%')
 );
 ?>
 
-
 <?php echo $this->Html->table($results, $tSettings);?>
-
 
 <script>
     $("#toggle-filter_fieldset").click(function () {
